client: Rename validate_tokens to tokens
